Prediction of Quantiles for Parametric Lifetime Distributions

This function predicts the quantiles of a parametric lifetime distribution using the (log-)location-scale parameterization.

predict_quantile(
  p,
  dist_params,
  distribution = c("weibull", "lognormal", "loglogistic", "sev", "normal", "logistic",
    "weibull3", "lognormal3", "loglogistic3", "exponential", "exponential2")
)

Arguments

p

A numeric vector of probabilities.

dist_params

A vector of parameters. An overview of the distribution-specific parameters can be found in section 'Distributions'.

distribution

Supposed distribution of the random variable.

Value

A vector with predicted quantiles.

Details

For a given set of parameters and specified probabilities the quantiles of the chosen model are determined.

Distributions

The following table summarizes the available distributions and their parameters

location parameter \(\mu\),

scale parameter \(\sigma\) or \(\theta\) and

threshold parameter \(\gamma\).

The order within dist_params is given in the table header.

distribution	dist_params[1]	dist_params[2]	dist_params[3]

"sev"	\(\mu\)	\(\sigma\)	-

"weibull"	\(\mu\)	\(\sigma\)	-

"weibull3"	\(\mu\)	\(\sigma\)	\(\gamma\)

"normal"	\(\mu\)	\(\sigma\)	-

"lognormal"	\(\mu\)	\(\sigma\)	-

"lognormal3"	\(\mu\)	\(\sigma\)	\(\gamma\)

"logistic"	\(\mu\)	\(\sigma\)	-

"loglogistic"	\(\mu\)	\(\sigma\)	-

"loglogistic3"	\(\mu\)	\(\sigma\)	\(\gamma\)

"exponential"	\(\theta\)	-	-

"exponential2"	\(\theta\)	\(\gamma\)	-

Examples

# Example 1 - Predicted quantiles for a two-parameter weibull distribution:
quants_weib2 <- predict_quantile(
  p = c(0.01, 0.1, 0.5),
  dist_params = c(5, 0.5),
  distribution = "weibull"
)

# Example 2 - Predicted quantiles for a three-parameter weibull distribution:
quants_weib3 <- predict_quantile(
  p = c(0.01, 0.1, 0.5),
  dist_params = c(5, 0.5, 10),
  distribution = "weibull3"
)