Beta Binomial Confidence Bounds for Quantiles and Probabilities

This non-parametric approach computes the beta binomial bounds (BB) for quantiles and failure probabilities using a procedure similar to the calculation of probabilities in terms of (Median Ranks).

confint_betabinom(x, ...)

# S3 method for model_estimation
confint_betabinom(
  x,
  b_lives = c(0.01, 0.1, 0.5),
  bounds = c("two_sided", "lower", "upper"),
  conf_level = 0.95,
  direction = c("y", "x"),
  ...
)

# S3 method for model_estimation_list
confint_betabinom(
  x,
  b_lives = c(0.01, 0.1, 0.5),
  bounds = c("two_sided", "lower", "upper"),
  conf_level = 0.95,
  direction = c("y", "x"),
  ...
)

Arguments

x
Object of class model_estimation (or model_estimation_list) returned from rank_regression.

b_lives
A numeric vector indicating the probabilities p of the B_p lives (quantiles) to be considered.

bounds
A character string specifying of which bounds have to be computed. One of "two_sided", "lower" or "upper".

conf_level
Confidence level of the interval.

direction
A character string specifying the direction of the confidence interval. One of "y" (failure probabilities) or "x" (quantiles).

Value

A tibble with class "confint" containing the following columns:

x : An ordered sequence of the lifetime characteristic regarding the failed units, starting at min(x) and ending up at max(x). With b_lives = c(0.01, 0.1, 0.5) the 1%, 10% and 50% quantiles are additionally included in x, but only if the specified probabilities are in the range of the estimated probabilities.

rank : Interpolated ranks as a function of probabilities, computed with the converted approximation formula of Benard.

prob : An ordered sequence of probabilities with specified b_lives included.

lower_bound : Provided, if bounds is one of "two_sided" or "lower". Lower limit of the confidence region with respect to direction, i.e. quantiles or probabilities.

upper_bound : Provided, if bounds is one of "two_sided" or "upper". Upper limit of the confidence region with respect to direction, i.e. quantiles or probabilities.

distribution : Specified distribution (determined when calling rank_regression).

bounds : Specified bound(s).

direction : Specified direction.

method : Specified method for the estimation of failure probabilities (determined when calling estimate_cdf).

Details

The difference to Median Ranks, i.e. finding the probability of rank j at a 50% level, is to determine the probability of rank j on another level, the specified confidence level.

See also

Examples

# Reliability data preparation:
## Data for two-parametric model:
data_2p <- reliability_data(
  data = shock,
  x = distance,
  status = status
)

## Data for three-parametric model:
data_3p <- reliability_data(
  data = alloy,
  x = cycles,
  status = status
)

# Probability estimation:
prob_tbl_2p <- estimate_cdf(
  x = data_2p,
  methods = "johnson"
)

prob_tbl_3p <- estimate_cdf(
  x = data_3p,
  methods = "johnson"
)

prob_tbl_mult <- estimate_cdf(
  x = data_3p,
  methods = c("johnson", "mr")
)

#> The 'mr' method only considers failed units (status == 1) and does not retain intact units (status == 0).

# Model estimation with rank_regression():
rr_2p <- rank_regression(
  x = prob_tbl_2p,
  distribution = "weibull"
)

rr_3p <- rank_regression(
  x = prob_tbl_3p,
  distribution = "lognormal3",
  conf_level = 0.90
)

rr_lists <- rank_regression(
  x = prob_tbl_mult,
  distribution = "loglogistic3",
  conf_level = 0.90
)

# Example 1 - Two-sided 95% confidence interval for probabilities ('y'):
conf_betabin_1 <- confint_betabinom(
  x = rr_2p,
  bounds = "two_sided",
  conf_level = 0.95,
  direction = "y"
)

# Example 2 - One-sided lower/upper 90% confidence interval for quantiles ('x'):
conf_betabin_2_1 <- confint_betabinom(
  x = rr_2p,
  bounds = "lower",
  conf_level = 0.90,
  direction = "x"
)

conf_betabin_2_2 <- confint_betabinom(
  x = rr_2p,
  bounds = "upper",
  conf_level = 0.90,
  direction = "x"
)

# Example 3 - Two-sided 90% confidence intervals for both directions using
# a three-parametric model:

conf_betabin_3_1 <- confint_betabinom(
  x = rr_3p,
  bounds = "two_sided",
  conf_level = 0.90,
  direction = "y"
)

conf_betabin_3_2 <- confint_betabinom(
  x = rr_3p,
  bounds = "two_sided",
  conf_level = 0.90,
  direction = "x"
)

# Example 4 - Confidence intervals if multiple methods in estimate_cdf, i.e.
# "johnson" and "mr", were specified:

conf_betabin_4 <- confint_betabinom(
  x = rr_lists,
  bounds = "two_sided",
  conf_level = 0.99,
  direction = "y"
)