Estimation of Failure Probabilities using Median Ranks

mr_method() is no longer under active development, switching to estimate_cdf is recommended.

mr_method(
  x,
  status = rep(1, length(x)),
  id = NULL,
  method = c("benard", "invbeta"),
  ties.method = c("max", "min", "average")
)

Arguments

x

A numeric vector which consists of lifetime data. Lifetime data could be every characteristic influencing the reliability of a product, e.g. operating time (days/months in service), mileage (km, miles), load cycles.

status

A vector of ones indicating that every unit has failed.

id

A vector for the identification of every unit. Default is NULL.

method

Method for the estimation of the cdf. Can be "benard" (default) or "invbeta".

ties.method

A character string specifying how ties are treated, default is "max".

Value

A tibble with only failed units containing the following columns:

id : Identification for every unit.

x : Lifetime characteristic.

status : Status of failed units (always 1).

rank : Assigned ranks.

prob : Estimated failure probabilities.

cdf_estimation_method : Specified method for the estimation of failure probabilities (always 'mr').

Details

This non-parametric approach (Median Ranks) is used to estimate the failure probabilities in terms of complete data. Two methods are available to estimate the cumulative distribution function F(t):

"benard" : Benard's approximation for Median Ranks.

"invbeta" : Exact Median Ranks using the inverse beta distribution.

Examples

# Vectors:
obs   <- seq(10000, 100000, 10000)
state <- rep(1, length(obs))
uic   <- c("3435", "1203", "958X", "XX71", "abcd", "tz46",
           "fl29", "AX23", "Uy12", "kl1a")

# Example 1 - Benard's approximation:
tbl_mr <- mr_method(
  x = obs,
  status = state,
  id = uic,
  method = "benard"
)
#> Warning: `mr_method()` was deprecated in weibulltools 2.0.0.
#> Please use `estimate_cdf()` instead.

# Example 2 - Inverse beta distribution:
tbl_mr_invbeta <- mr_method(
  x = obs,
  status = state,
  method = "invbeta"
)
#> Warning: `mr_method()` was deprecated in weibulltools 2.0.0.
#> Please use `estimate_cdf()` instead.